refactor: remove module dependency in fedimint-lightning
