+++ gateway/fedimint-lightning/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod common;
 pub mod ldk;
 pub mod lnd;
 
@@ -9,19 +10,19 @@ use async_trait::async_trait;
 use bitcoin::address::NetworkUnchecked;
 use bitcoin::hashes::sha256;
 use bitcoin::{Address, Network};
+use common::{PrunedInvoice, RouteHint};
 use fedimint_core::encoding::{Decodable, Encodable};
 use fedimint_core::secp256k1::PublicKey;
 use fedimint_core::task::TaskGroup;
 use fedimint_core::util::{backoff_util, retry};
 use fedimint_core::{secp256k1, Amount, BitcoinAmountOrAll};
-pub use fedimint_ln_common::contracts::Preimage;
-use fedimint_ln_common::route_hints::RouteHint;
-use fedimint_ln_common::PrunedInvoice;
 use futures::stream::BoxStream;
 use lightning_invoice::Bolt11Invoice;
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
 use tracing::{info, warn};
+
+use crate::common::Preimage;
 
 pub const MAX_LIGHTNING_RETRIES: u32 = 10;
