+++ misc/git-hooks/pre-commit
@@ -89,7 +89,7 @@ function check_check_forbidden_dependencies() {
       return 1
     fi
     find modules/ -name Cargo.toml | grep common/ | while read -r cargo_toml ; do
-      if grep -E "fedimint-" "$cargo_toml" | grep -E -v "fedimint-core|fedimint-api-client|-common|fedimint-logging" >&2 ; then
+      if grep -E "fedimint-" "$cargo_toml" | grep -E -v "fedimint-core|fedimint-api-client|-common|fedimint-logging|fedimint-lightning" >&2 ; then
         >&2 echo "Fedimint modules' -common crates should not introduce new fedimint dependencies: $cargo_toml"
         >&2 echo "The goal is to avoid circular deps that blow up build times. Ping @dpc for help."
         return 1
