+++ modules/fedimint-gw-client/src/complete.rs
@@ -4,9 +4,9 @@ use fedimint_client::sm::{ClientSMDatabaseTransaction, State, StateTransition};
 use fedimint_client::DynGlobalClientContext;
 use fedimint_core::core::OperationId;
 use fedimint_core::encoding::{Decodable, Encodable};
+use fedimint_lightning::common::Preimage;
 use fedimint_lightning::{InterceptPaymentResponse, PaymentAction};
 use fedimint_ln_client::incoming::IncomingSmStates;
-use fedimint_ln_common::contracts::Preimage;
 use futures::StreamExt;
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
