+++ modules/fedimint-gw-client/src/pay.rs
@@ -10,12 +10,13 @@ use fedimint_core::config::FederationId;
 use fedimint_core::core::OperationId;
 use fedimint_core::encoding::{Decodable, Encodable};
 use fedimint_core::{secp256k1, Amount, OutPoint, TransactionId};
+use fedimint_lightning::common::Preimage;
 use fedimint_lightning::{LightningRpcError, PayInvoiceResponse};
 use fedimint_ln_client::api::LnFederationApi;
 use fedimint_ln_client::pay::{PayInvoicePayload, PaymentData};
 use fedimint_ln_common::config::FeeToAmount;
 use fedimint_ln_common::contracts::outgoing::OutgoingContractAccount;
-use fedimint_ln_common::contracts::{ContractId, FundedContract, IdentifiableContract, Preimage};
+use fedimint_ln_common::contracts::{ContractId, FundedContract, IdentifiableContract};
 use fedimint_ln_common::{LightningInput, LightningOutput};
 use futures::future;
 use lightning_invoice::RoutingFees;
