+++ modules/fedimint-ln-client/src/incoming.rs
@@ -18,8 +18,9 @@ use fedimint_core::core::OperationId;
 use fedimint_core::encoding::{Decodable, Encodable};
 use fedimint_core::runtime::sleep;
 use fedimint_core::{Amount, OutPoint, TransactionId};
+use fedimint_lightning::common::Preimage;
 use fedimint_ln_common::contracts::incoming::IncomingContractAccount;
-use fedimint_ln_common::contracts::{ContractId, Preimage};
+use fedimint_ln_common::contracts::ContractId;
 use fedimint_ln_common::LightningInput;
 use lightning_invoice::Bolt11Invoice;
 use serde::{Deserialize, Serialize};
