+++ modules/fedimint-ln-client/src/pay.rs
@@ -10,10 +10,10 @@ use fedimint_core::encoding::{Decodable, Encodable};
 use fedimint_core::task::sleep;
 use fedimint_core::time::duration_since_epoch;
 use fedimint_core::{secp256k1, Amount, OutPoint, TransactionId};
+use fedimint_lightning::common::{PrunedInvoice, RouteHint};
 use fedimint_ln_common::contracts::outgoing::OutgoingContractData;
 use fedimint_ln_common::contracts::{ContractId, FundedContract, IdentifiableContract};
-use fedimint_ln_common::route_hints::RouteHint;
-use fedimint_ln_common::{LightningGateway, LightningInput, PrunedInvoice};
+use fedimint_ln_common::{LightningGateway, LightningInput};
 use futures::future::pending;
 use lightning_invoice::Bolt11Invoice;
 use reqwest::StatusCode;
