+++ modules/fedimint-ln-common/src/contracts/outgoing.rs
@@ -2,9 +2,9 @@ use bitcoin::hashes::Hash as BitcoinHash;
 use fedimint_core::encoding::{Decodable, Encodable};
 use fedimint_core::secp256k1::PublicKey;
 use fedimint_core::Amount;
+use fedimint_lightning::common::Preimage;
 use serde::{Deserialize, Serialize};
 
-use super::Preimage;
 use crate::contracts::{ContractId, IdentifiableContract};
 use crate::LightningInput;
