+++ modules/fedimint-ln-tests/Cargo.toml
@@ -21,6 +21,7 @@ fedimint-core = { workspace = true }
 fedimint-dummy-client = { workspace = true }
 fedimint-dummy-common = { workspace = true }
 fedimint-dummy-server = { workspace = true }
+fedimint-lightning = { workspace = true }
 fedimint-ln-client = { workspace = true }
 fedimint-ln-common = { workspace = true }
 fedimint-ln-server = { workspace = true }
